requirements.txt
fastmcp
beautifulsoup4
lxml
PyYAML
jsonref
